remaining init exps

+++ model/ctn.py
@@ -119,7 +119,7 @@ class Net(torch.nn.Module):
         self.kl = nn.KLDivLoss()
         self.samples_seen = 0
         self.samples_per_task = self.cfg.samples_per_task
-        self.sz = self.cfg.replay_batch_size
+        self.sz = int(self.cfg.replay_batch_size)
         self.inner_steps = self.cfg.ctn_inner_steps
         self.n_meta = self.cfg.ctn_n_meta
         self.count = 0
@@ -174,8 +174,8 @@ class Net(torch.nn.Module):
             mem_x = self.memx[:t,:]
             mem_y = self.memy[:t,:]
             mem_feat = self.mem_feat[:t,:]
-            sz = min(self.n_memories, self.sz)
-            idx = np.random.choice(t* self.n_memories,sz, False)
+            sz = int(min(self.n_memories, self.sz))
+            idx = np.random.choice(int(t * self.n_memories), sz, False)
             t_idx = torch.from_numpy(idx // self.n_memories)
             s_idx = torch.from_numpy( idx % self.n_memories)
             offsets = torch.tensor([self.compute_offsets(i) for i in t_idx]).cuda()
